Add a debian/gitlab-ci.yml file to the debian-sid branch
It seems as if gitlab can only copy with a single CI configuration
path per repository, not per branch.

Since the master and the debian-sid branch need different Gitlab CI
configurations and the debian-sid branch must not overwrite files from
the master branch, this is difficult to reach. The only albeit
suboptimal solution would be to add a debian/gitlab-ci.yml file to the
master branch even though it has no debian directory (yet) as the
upstream debian directory is always ignored upon orig.tar import.

So for now, we only have CI in the debian-sid branch.