Build-Depend on libfmt-dev to be able to use it in the code, later std::format
This library has been used as base and then co-developed with the
standardisation process of std::format of the C++20 standard, approved a few
months ago.

So by converting code using Boost::Format and others like snprintf(), the code
will be ready to pivot to std::format when it becomes available in the default
Debian toolchain.